+++ src/main/java/ru/yandex/practicum/filmorate/mapper/DirectorFromFilmMapper.java
@@ -3,19 +3,14 @@ package ru.yandex.practicum.filmorate.mapper;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.ResultSetExtractor;
 import ru.yandex.practicum.filmorate.model.Director;
+import java.sql.*;
+import java.util.*;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-
-public class DirectorFromFilmMapper implements ResultSetExtractor<Map<Integer, List<Director>>> {
-    private final Map<Integer, List<Director>> directorsOfFilms = new HashMap<>();
+public class DirectorFromFilmMapper implements ResultSetExtractor<Map<Long, List<Director>>> {
+    private final Map<Long, List<Director>> directorsOfFilms = new HashMap<>();
 
     @Override
-    public Map<Integer, List<Director>> extractData(ResultSet rs) throws SQLException, DataAccessException {
+    public Map<Long, List<Director>> extractData(ResultSet rs) throws SQLException, DataAccessException {
         while (rs.next()) {
             if (isFilmDirectorsInMap(rs)) {
                 addToExistingList(rs);
@@ -27,18 +22,18 @@ public class DirectorFromFilmMapper implements ResultSetExtractor<Map<Integer, L
     }
 
     private boolean isFilmDirectorsInMap(ResultSet rs) throws SQLException {
-        return directorsOfFilms.containsKey(rs.getInt("film_id"));
+        return directorsOfFilms.containsKey(rs.getLong("film_id"));
     }
 
     private void addToExistingList(ResultSet rs) throws SQLException {
-        List<Director> directors = directorsOfFilms.get(rs.getInt("film_id"));
+        List<Director> directors = directorsOfFilms.get(rs.getLong("film_id"));
         directors.add(new Director(rs.getInt("director_id"), rs.getString("name")));
-        directorsOfFilms.put(rs.getInt("film_id"), directors);
+        directorsOfFilms.put(rs.getLong("film_id"), directors);
     }
 
     private void addNewList(ResultSet rs) throws SQLException {
         List<Director> directorsOfOneFilm = new LinkedList<>();
         directorsOfOneFilm.add(new Director(rs.getInt("director_id"), rs.getString("name")));
-        directorsOfFilms.put(rs.getInt("film_id"), directorsOfOneFilm);
+        directorsOfFilms.put(rs.getLong("film_id"), directorsOfOneFilm);
     }
 }
