+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -1,24 +1,17 @@
 package ru.yandex.practicum.filmorate.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
+import lombok.*;
 import ru.yandex.practicum.filmorate.validator.date.AfterMinDate;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.Size;
+import javax.validation.constraints.*;
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 @Data
 @RequiredArgsConstructor
 @AllArgsConstructor
 public class Film {
-    private int id;
+    private Long id;
     @NotBlank
     private String name;
     @NotBlank
@@ -30,13 +23,13 @@ public class Film {
     private LocalDate releaseDate;
     @NotNull
     @Positive
-    private int duration;
+    private Integer duration;
     @NotNull
     private RatingMpa mpa;
     private List<Genre> genres = new ArrayList<>();
     private List<Director> directors = new ArrayList<>();
 
-    public Film(String name, String description, LocalDate releaseDate, int duration,
+    public Film(String name, String description, LocalDate releaseDate, Integer duration,
                 RatingMpa mpa, List<Genre> genres, List<Director> directors) {
         this.name = name;
         this.description = description;
